[TEST] Make sure random index template uses the proper number of data nodes in bw comp tests
When creating indices with no specific number of replicas set, the random index template is used, which has a maximum number of replicas set to the total number of data nodes - 1.
That works well in usual tests, but not for bw comp tests as we have an allocation decider that prevents allocation of replicas when the primary is allocated on a newer node. That means that replicas are not allocated and in some cases there's no quorum and index operations time out.