[TEST] dynamic scripts to be enabled on demand
Back when tests ran against the global cluster and we disabled groovy dynamic scripts by default, we made it easy to use scripts in every test.Now that global cluster is gone and the default is suite, we should disable dynamic scripts by default, same behaviour as our released artifacts. Let the different tests that need scripting enabled it on demand instead.