Introduce RestAsyncClient that uses apache async http client
We now have an abstract base class called AbstractRestClient that is used to share common code between the sync and the async client. The two clients effectively expose a different performRequest method.

RestClient stays the original sync client, RestAsyncClient accepts a listener that will be notified when the request is successfully completed or failed. It also allows to stream the response through an HttpAsyncResponseConsumer.