Increase warmup iterations for index benchmark
After analyzing JIT outout it was clear that 10% warmup samples
are too little for the bulk index benchmark. The JIT was still
actively compiling code at this point in time.

With this commit we increase the number of warmup iterations
to 40% of all bulk requests and introduce additional documentation
on useful startup options for the benchmark.