Add support for runtime fields
This commit includes the work that has been done on the runtime fields feature branch to add support for runtime fields added to index mappings. The high level tasks are listed in #59332.

We are adding a new x-pack plugin called runtime-fields that plugs in a custom mapper which allows to define runtime fields based on a script.
The changes included in this PR that were made outside of the x-pack/plugin/runtime-fields directory are minimal and revolve around 1) making the ScriptService available while parsing index mappings so that the scripts associated to runtime fields can be compiled 2) sharing code to manipulate ranges etc. as it can be reused in runtime fields .

Co-authored-by: Nik Everett <nik9000@gmail.com>