Remove timeout handling from fetch phase
This is mostly a revert of #116676 .
While we want to handle search cancellations in the fetch phase, and
gracefully stop the execution when the search is cancelled during fetch,
we should not handle search timeouts in the fetch phase.

The reason for that is that there are no partial top hits results returned
if the fetch phase times out, hence running uninterrupted fetch phase after a
query phase timeout is necessary to provide partial top hits.

In case the query phase does not time out, it also feels weird that some top hits
may be skipped because fetch times out on some shard. This scenario is not handled
in the coord node at all and causes an array index out of bounds exception currently.

It may make sense to honour the timeout in the fetch phase when partial results are
disallowed, but that is counter intuitive in that it introduces differential timeout
behaviour depending on whether partial results are allowed or not.

Closes #130071